in theory only select LaTeX source of noneditables
in practice, only works in Firefox.
In WebKit, things that are user-select:none still get selected and copied when you do Ctrl+C, they just aren't visually highlighted. IE and Opera do not support user-select.
The "unselectable" attribute usually recommended as the IE and Opera equivalent actually isn't. It is equivalent to returning false onselectstart, and doesn't affect selections that start elsewhere, like dragging from somewhere else over the math or Ctrl+A. It's actually kind of similar to user-select's behavior in WebKit.