added \underline and \bar = \overline
technically in real LaTeX \bar isn't the same as \overline, it's an inferior precursor to \overline. It only shows a short horizontal line the width of a single letter above the middle of the contents whereas overline shows up over the entire contents, which is dumb, and the only reason you'd use \bar is because it's shorter to type and looks the same for single-letter variables, and the only reason it's still there is for backwards compatibility.