git push all tracking branches, not just gh-pages
This way, if you commit to a remote-tracking branch but don't push it, and then ./build.sh publish it, the script will push your other commit in addition to the new gh-pages commit, so when it says "publish latest 123abcd commit message" the word 123abcd will link to your commit.

This also means you can't (well, shouldn't) rebase or amend after publishing.