first attempt at \widehat, \over{left|right}arrow
despite my best efforts, still does not quite
work in IE <9:
- tweaks to matrix filter fallback for scale
broke sqrt and paren
- when selecting with the mouse, blocks of blue
highlight color will flash and disappear
(explained below)

So, barring painstakingly calculating the exact
dimensions of what the font-size-scaled-up-and-
then-matrix-shrunk widehat character is, I can't
guarantee that no background will go over anything
else (such as the contents below the hat), so I
have to use the Chroma filter to make the
background transparent (an opaque background is
necessary for the matrix filter to anti-alias
properly, see the Wiki page "Transforms").
However, while multiple CSS filters on the same
element are supported, so both a Chroma filter
and matrix filter in the same CSS rule would work,
the background color changes on selection which
only affects the CSS descendant selectors and
triggers no JS hooks, which can happen between
redraws when the matrix filter is calculated, and
a different CSS rule for the Chroma filter (to
match the changed background color) would override
the matrix filter that shrinks it the exact right
calculated amount. "Luckily" the Chroma filter
works even on contents, so we can set the Chroma
filter on a parent element that will change with
the changing background color depending on
selection/blur/etc, while the matrix filter on the
child element stays the same. (Note: that actually
the Chroma filter doesn't work with system color
keyword for selection highlight color, so I only
use the default Windows highlight color, and if
they have a custom color, there will be a fringe
of highlight text color anti-aliased against the
background of the default Windows highlight color,
inside what otherwise appears to be the custom
color. See the comment in the stylesheet for a
more thorough explanation.) Apparently IE has
repaint problems with filters, though, because
when selecting with the mouse, the large blocks
with solid background color equal to the selection
highlight color but chroma keyed to be transparent
become momentarily opaque, until the filter kicks
in and the large block of color becomes
transparent again. This can probably be solved by
using a less dumb mouse selection algorithm;
currently, every mousemove, it will repeatedly
selectLeft or Right from where you initially
mousedowned to where your mouse currently is.

Also, aliases in Chrome on Windows. As far as I
can tell, this is unfixable:
http://www.quora.com/Cascading-Style-Sheets/Why-does-stretching-text-horizontally-with-a-CSS-transform-alias-in-Chrome-on-Windows)