don't depend on dom-level selection apis as much
Before merging, we need to test this thoroughly in IE8.  This prevents
handleText() from being called erroneously on keydowns, but keypresses
will still trigger a call.  According to http://unixpapa.com/js/key.html
this should be fine, because these IE < 9 does not trigger keypress
events for special keys.  But we should still test it first :).