fix LatexCommandInput prepend bug
fix long-standing bug that if you go left until you're outside the
LatexCommandInput block, and then you go right again, the cursor will be
to the left of the \ and things you can type there. Similarly, if you
type anything at all so that the LatexCommandInput is nonempty, and then
hit Up, the cursor will be to the left of the \.

Both of these bugs are because Cursor::prependTo calls jQuery::prependTo
on the block's jQ, so it can only contain math contents and extra stuff
has to go outside.

The reason the long-standing bug was there was that by having an inner
span without the \, when the LatexCommandInput is focused, the inner
span without the \ gets highlighted with the hasCursor CSS class and the
backslash is outside it, which looks wrong. The fix is to monkey-patch
focus and blur on the block to add and remove the hasCursor CSS class to
and from the outer span containing both the \ and the block span. This
is ok because LatexCommandInput is one big hack anyway.