attribute macros are implicit in new HTML template
The diff is big but the changes are small:
- change baseclasses.js comments to describe new implicit HTML template
  format
- change MathCmd::html() to be able to parse it
- unit tests for MathCmd::html()
- Symbol::init no longer needs to add #mqCmdId attribute macro,
  because MathCmd::html() is now a generalization thereof
- :%s/ #mqCmdId//g, :%s/ #mqBlockId:\d\+//g, :%s/#mqBlock:/#/g

With rules for placing the attribute macros so mechanical, it clearly
should be automated. The implicit HTML template is also way prettier,
at the cost of a somewhat more complex MathCmd::html() method, a
stricter requirement on the well-formedness of the HTML (which all of
our HTML templates already conform to), and a stricter requirement on
the placement of block content markers, but the only violator was
LatexCommandInput which was fixed in the previous commit, and the
requirement is necessary for other reasons, which had resulted in a
long-standing bug in LatexCommandInput that was fixed in the previous
commit.