Feature Change: backspace/delFwd select empty cmds
the previous behavior was, when and only when backspacing/deleting
nonempty commands, like a nonempty fraction or a nonempty sqrt, to
select the cmd, otherwise directly remove it. The new behavior is to
only directly remove symbols, all commands that may have blocks will be
selected, including empty ones. This is more internally consistent.

(Prep for deletion being an inherited method from MathCommand that
Symbol will override)