RootMathBlock::onKey now calls Cursor::escapeDir
which delegates to this.parent.moveOutOf

per the Occasionally Useful Suggestion of Demeter, it's really not the
root block's business to be looking at the cursor's parent and next and
telling the cursor where to go. In fact, it duplicated code from
Cursor::moveDirWithin that relied on the double-layered-ness of the
tree.

To sum up, in one semantic change this improves the code in 3 ways:
1. refactors duplicate code between RootMathBlock::onKey and
   Cursor::moveWithinDir
2. moves cursor logic into the cursor from the root editable
3. moves tree logic into the tree node from the cursor/root editable