refactor generator-iterator pattern into util HOF
Node::bubble, Node::postOrder, and Fragment::each are all implicit
iterator functions that repeatedly call their argument (as a function or
as a method name) on each of some set of objects.

This abstracts that away, letting them be defined in terms of what they
do differently, as "generators" that "yield" objects repeatedly.