fix #120 TextBlocks couldn't be created
This was because when we switched MathCommand::createBlocks from munging
pointers to calling Node::adopt in 3d9ec92, there was a bug where we
accidentally removed the .blur() call on all the blocks, which was fixed
in caab952 by calling .blur() on all the blocks in
MathCommand::createBefore, after the createBlocks call. Unbeknownst to
us, TextBlock overrides createBlocks not just to create an
InnerTextBlock rather than a MathBlock as it's child, but also
intentionally didn't call .blur() (which removes the entire TextBlock,
since it thinks it's being called because the cursor left the TextBlock
and the TextBlock is empty).