LatexCommandInput hooks onText instead of write()
it now monkey-patches its child block, overriding MathBlock::write(),
rather than overriding onText on itself. Aside from the monkey-patching
(which will go away later in this environments refactoring as we get rid
of the double-layered tree), this is much cleaner because the cursor
implicitly takes care of cursor stuff for us, in fact it just naturally
fixes #123.