fix renderLatex() with toolbar
... i.e., in the mathquill-editor, by, rather than assuming that the
textarea is the first thing in the root jQ and that it is the only thing
we want to preserve when emptying the textarea, instead assume that the
toolbar, if present, must precede the textarea.

Note that the only way there could be no this.textarea is on initial
render, in which case there's no need to clear the RootMathBlock
anyway.