'onKey' event bubbles with cursor as an argument
much cleaner than storing it as this.cursor on createLeftOf, like
LatexCommandInput and Vector were doing.

Note that the entire line of the this.cursor assignment in
LatexCommandInput::createLeftOf could be removed because the
insAtRightEnd was unnecessary, placeCursor already does that.

This also fixes \vector when rendered from LaTeX (as if we give a shit,
that shit should've been on a experimental branch anyway)