Separate out root block DOM node from container DOM node #optimize #scary
We can get a perf bump by not making the redundant DOM calls to detach
the textarea and prepend it again, by making the textarea span and the
root block span siblings, rather than the textarea span always being the
first child of the root block span.

Typical profiling log results:
https://gist.github.com/laughinghan/5938626

This is kind of a scary change because the root node code is kind of
messy, I was way surprised how easy it was.

Turns out to also be much more elegant: The root block used to be the
only exception to the invariant that a block's DOM node has as its
children exactly the DOM nodes of the block's children in the edit tree.
The invariant being universally true removes an ugly special case from
Cursor::insAtDirEnd.