Fix paste in inner editable math fields
For editables the paste handler is attached to the container so as to
listen for middle-click paste, but for static math the only reason for
the cut and paste handlers is that Ctrl-X and Ctrl-V don't do anything
for static text so when static math is selected, even though a textarea
is actually selected, we want to pretend they don't do anything. So
there's no reason to listen for cut or paste events from anywhere beside
the textarea, and in fact preventing default on any paste event from
anywhere inside the static math blocks paste in inner editable math
fields.