Stop triggering spurious focusin events on click
Only focus textarea if it became unfocused (which it only should in IE<9
where preventing default in mousedown fails to prevent focus from
transferring to the mousedown target (which when it's anything in a
MathQuill span, we'd like to prevent)).

New contract on textarea.focused: set in focus/blur event handlers on
the textarea, elsewhere it'll be correct when checked.