Replace `textarea.focused` with `root.blurred`
`textarea` is a jQuery object, shouldn't put flag on there, whereas
`root` is a long-lived RootMathBlock.

Grepping for `.focused`, literally every condition on it negated it, so
use `.blurred` instead. (The name is also shorter.) The downside is the
initial value has to be set to true (rather than defaulting to the falsy
`undefined`), which it already was for editables because of the initial
blur on the textarea (in `focusBlurEvents()`), but wasn't for static
math.