Fix static math blur/timeout/select brittleness
This change should have been made when textarea-detaching/reattaching
was moved out of `root.selectionChanged()`, line 99/95 of 0841d9f:
https://github.com/mathquill/mathquill/commit/0841d9fd24d92523e19673492f0708ea246dae46#diff-6fcb8bd3232471c30d0ec6b0299630a0L99

It wasn't, and should have resulted in a bug when, in a later commit,
`textarea.select()` was moved into a timeout:
https://github.com/mathquill/mathquill/commit/e0ceef69cc6ce31dc6478635650caeaf21b2158f#commitcomment-5219665

Basically, it should have resulted in a bug where, after selecting some
static math, clicking on something else focusable (e.g. another
MathQuill, or a plain textarea) would fail to focus the other focusable
thing, because the static math's textarea would "steal back" focus just
before being detached:

1. after selecting some static math, the static math's textarea is
   focused,
2. then when something else focusable is clicked, a `blur` event fires
   on the static math's textarea,
3. which calls `cursor.clearSelection()` to clear the (typically, blue)
   selection highlight, but which **also**
4. calls `root.selectionChanged()`,
5. which, after a timeout (so after the other focusable thing has become
   focused), selects the static math's textarea, "stealing back" focus.

The only reason this bug didn't surface was that
a962f1fe1ba0785f54e0460c525fbcb0b833f3fe happened to make it so that
(what has now become) `setTextareaSelection()` doesn't call
`textarea.select()` unless its contents are nonempty, probably with the
thought that selecting an empty textarea is just an expensive noop, but
that's only true when the textarea is focused; in this case, it masks
this bug, because when the textarea is blurred, `textarea.select()`
would focus the textarea, "stealing" focus away from the most recently
focused element.