Move onNotify 'edit' from Cursor to Controller
Next to Controller::deleteDir, its main usage. (Also used in
Cursor::writeLatex and the cut event handler in textarea.js, which are
kinda secondary to the 'Backspace' key, I feel like. Really though, I
want to move the notify stuff from Cursor to Controller, and this is the
last onNotify() call in cursor.js.)