Make selectionChanged a normal Controller method
...rather tha monkey-patched in setSelectionChangedFn.

.selectFn still has to be monkey-patched, but just do that in
Controller::staticMathTextareaEvents and editablesTextareaEvents.

Note that setTextareaSelection must be an exposed method and can't just
be an anonymous function passed to setTimeout() because it also needs to
be called synchronously (not in a timeout) in the 'copy' event handler.

In the 'focus' event handler, ctrlr.selectionChanged() waits a timeout
before calling .select() on the textarea anyway, no reason to wait
another timeout (plus would have to make an anonymous function, because
it's now a normal method that uses `this`).