Tweak saneKeyboardEvents() contract
Callbacks renamed from .key() and .text() to the more descriptive
.keystroke() and .typedText(); they are also called directly on the
object passed in to preserve `this`. (Prep for passing in the controller
object directly as the `handlers` object.)

The object and all the handlers on it are now required. This means that
saneKeyboardEvents.test.js no longer needs shouldNotBeCalled(), just
don't pass in handlers for the ones that should never be called. This
also means handlers that may be called but should be ignored have to
explicitly be set to noops now.