Auto-unitalicize sin, log etc as seq of Variable's #big
Rather than each being a single NonItalicizedFunction command like \sin,
they're now a sequence of individual letters (Variable's) with special
styling. Any sequence of Variable's 's', 'i', and 'n' will be
automatically unitalicized to look like \sin, but the cursor can go
between and edit individual letters.

Node::respace is the only hook called when a sibling, rather than a
descendant, is inserted or deleted, so Variable::respace is where we
have to search for the relevant contiguous run of letters and
unitalicize any substrings therein.

Also add an UnItalicized class so that typing or parsing \sin, \cos,
\log etc results in the appropriate sequence of Variable's.