Parens now have a ghost that auto-expands to end #big
When brackets are now typed, they come with a matching "ghost" bracket
at the far end of the current block. This is done by auto-expanding to
contain the rest of current block and being set to a "one-sided" state.

Then, typing a matching bracket facing the opposite way in a one-sided
bracket "closes" the matching bracket pair around anything between them.
This is done by moving everything up til the ghost outside the bracket
pair and "solidifying" the ghost.

Also, having to add {Node,Fragment}::withDirAdopt() was annoying.
Not sure what could be done about that; it's non-trivial to
directionalize Fragment::adopt, its implementation relies on the
order that .each() iterates through the nodes.