Parens/brackets/braces now deleted one at a time
You can now delete one of a pair of matching brackets, which turns into
a ghost and auto-expands to the far end of the current block, making the
pair one-sided again. Deleting the non-ghost side of a one-sided bracket
deletes it completely.

The .delete{Towards,OutOf}() architecture is so great! But not the
having to monkey-patch methods on the sole MathBlock in finalizeTree,
that sucks.