Fix deleting out of brackets when ghost already at end
In cases like 1+(|a+b), backspacing the open paren results in (1+|a+b),
with the cursor always next to what was the first child of the paren,
which is great. However, in cases like (|a+b), backspacing the open
paren results in (|a+b) still because there's nowhere for the ghost to
auto-expand to and then the cursor was still being inserted next to the
first child of the paren, when what we want is |(a+b) after backspacing
the open paren.