Reorganize Bracket's .delete{Toward,OutOf}() #reorganize
Factor out Bracket::deleteSide(), superseding Bracket::oneSideify(),
that takes care of the case of deleting the non-ghost side of a
one-sided bracket (where you want to unwrap the parens completely), and
also takes care of where the cursor should go.

This is prep for additional pairing up of one-sided brackets on deletion
(in the next commit).