Fix edge case where ghost might not be at end
As you can see in the test case, the edge case is if you have:

    1+((2+3))

where the inner open-paren is a ghost, and you delete the outer
open-paren, which becomes a ghost and auto-expands to the far left end:

    (1+(2+3))

Before, the inner open-paren would remain a ghost, just hanging out, not
at the end of its block. Now, we call .siblingCreated() on the inner
open-paren to solidify it.

Note that there's no need for a symmetric call to origEnd's new sibling
because it gets .siblingDeleted() called on it by Controller::deleteDir
(in keystrokes.js).

(These should probably be changed to just .siblingChanged(), seeing as
origEnd's new sibling wasn't just created.)