Handle intentional blur
That is, if the typist deliberately clicks outside an editable field,
the selection clears and any ghost parens solidify; but if the typist
just switches tabs/windows and switches back, nothing should change.

[QuirksMode][1] reports a few browser bugs with focus events on the
window but none with blur events on the window registered in the
bubbling phase.

[This widely cited blog post][2] agrees except that in IE6/7 (others
versions not mentioned), spurious blur events fire on the window "when
the focus went from the body to an input or link"; in other words, if
the page is in focus but no particular element is focused
(document.activeElement === document.body), and then something on that
page is focused (the page remains in focus, so sane browsers don't fire
blur on the window). Luckily that doesn't affect this code, we're not
listening for blur on the window except immediately after a blur on
MathQuill's hidden textarea.

[1]: http://www.quirksmode.org/dom/events/blurfocus.html
[2]: http://www.thefutureoftheweb.com/blog/detect-browser-window-focus