Change latex for ' ' from '\\:' to '\\ '
Writing a backslash (which triggers LatexCommandInput) and then hitting
Space, Tab, or Enter used to insert the backslash character when the
latex of the left end was empty when really it should have inserted a
space.

We changed `CharCmds[' ']` to a `LatexCmds[' ']` since latexMathParser
can parse `\ ` fine and `LatexCmds` only checks for the control
sequence.