Make Selection() take fixed number of arguments
The shortcut is only ever used in one place, and resulted in a
hard-to-debug infinite loop that could've been immediately debugged
if a prayer had failed instead:

When writing the previous commit, at first I forgot to change the
signature of Node::selectChildren() from (cursor, leftEnd, rightEnd) to
(leftEnd, rightEnd), but changed everything else. So,
Node::selectChildren() was being called with 2 arguments, leftEnd
had what was actually the right end of the selection and rightEnd was
undefined, so Selection()'s constructor was in one-argument mode and the
cursor's selection was always a one-node fragment of what should be the
right end node. EditableField::select, then, which does

    while (ctrlr.cursor[L]) ctrlr.selectLeft();

infinite looped, since the selection never changes.