Fix cursor position after deleting a SupSub block
In a 2-block SupSub (i.e. with a superscript stacked on top of a
subscript):

Before, backspacing out of a superscript (or deleting forward out of a
subscript) would put the cursor at the right end of the remaining
subscript (or at the left end of the remaining superscript), which is
unfortunate because it's not symmetrical with typing back the deleted
superscript (or subscript).

Now, backspacing out of a superscript (or deleting forward out of a
subscript) puts the cursor to the left of the now subscript-only SupSub
(or to the right of the now superscript-only SupSub), from where
subscripts and superscripts can be typed back.