Assuage #293 Cmd-Left turns seln into typed text
First, fix in all browsers but Firefox (and Opera maybe? Not fixed in
browsers where arrow keys fire keypress) by always re-selecting the
textarea, in case an unrecognized key like Cmd-Left clears the
selection.

As for Firefox, "we just assume that if the text is 1 unselected
character then it must have been typed (no matter whether different from
or same as selection), the only bug will be that when exactly 1
character is selected, unrecognized keys in Firefox will clear the
selection":
https://github.com/mathquill/mathquill/issues/293#issuecomment-40997668