Fix not being able to blur if there's a selection
Revert bad hunk that tries to always ensure selection text is always
selected, because in a timeout after blur, .select('') is called on the
shim, which then calls checkTextarea() one last time before munging the
textarea but which re-selects and hence re-focuses the textarea:
https://github.com/mathquill/mathquill/pull/294#issuecomment-41861819

Instead, select() just sets a flag:

    shouldBeSelected = !!text;

that onKeydown() checks to see if it should re-select the textarea,
which it only does once so that a subsequent .select('') after blur
won't re-focus the textarea.

Note: do this onKeydown() not handleKey() because the other caller of
handleKey(), onKeypress(), immediately thereafter calls
checkTextareaFor() and would overwrite checkTextarea.