Rename external .mathquill-{editable => math-field}
Styling should be only done with internal CSS classes, so now the README
refers to .mathquill-math-field and never .mathquill-editable (which
should be .mq- not .mathquill-).

The new .mathquill-math-field CSS class is entirely external, its only
mention internally is on document ready, find them all and render them
as MathQuill.MathField().

Recommend --word-diff-regex='[a-z-]+|.' to view diff.