Kill setMathQuillDot()'s hardcoded string keys #refactor #cleanup
This was dumb and smelled even when it was introduced:
https://github.com/mathquill/mathquill/pull/104#discussion-diff-1175704

Not only is it obviously simpler to just set MathQuill.MathField (etc)
than a variable key that's actually one of a few hardcoded strings, but
APIFnFor(APIClass) is such a more descriptive name and signature, like,
what did setMathQuillDot() set MathQuill dot to anyway? Creating the
corresponding public API function for an API class is what most of this
code is actually about.

Also rewrite relevant comment, and rewrite relevant README entry to
match phrasing of comment.