overrideAutoOperatorNames => autoOperatorNames option
Replace MathQuill.overrideAutoOperatorNames() with `autoOperatorNames`
config option.

The big external change is now there are no weird standalone "Global
Behavior Options", all options are config options and may be set
globally with `MathQuill.config({ ... })`, finishing resolving #286
(specifically the part about being able to remove auto-commands, set
them per-field, etc).

The big internal architectural change is that finalizeTree,
siblingCreated, and siblingDeleted now take in the options object as
their first argument, a change that has to happen in unfortunately many
places. Luckily, that change is completely straightforward in all but
one of those places.

The exception is MathElement::finalizeInsert, which appears to take in a
cursor as an argument, so I wanted to just use `cursor.options`, but it
turns out the `cursor` parameter is part of an opaque contract where
only Controller::writeLatex passes in a cursor (apparently) [0].
I'll add a comment about this weird contract in the next commit.

[0]: https://github.com/mathquill/mathquill/commit/ea7307eb4fac77c149a11ffdf9a831df85247693#commitcomment-7378712