Fix extra .scrollWidth on math field ancestor
Without this, the hidden textarea (which has browser default dimensions
of usually several em wide and a couple em tall) can "stick out"
invisibly from the right and bottom edges of a math field (which can be
seen by highlighting the textarea in Chrome's Web Inspector, for
example).

I don't think "sticking out" invisibly can have any other effect, but
apparently it can affect the `.scrollWidth` or `.scrollHeight` of the
nearest ancestor that isn't `overflow:visible` (or the viewport, of
course). If the textarea is sticks out of such an ancestor (because the
math field is closer to the ancestor's right or bottom edge than however
far the textarea sticks out), then like anything `position:absolute`
sticks out to the right or below, the ancestor's [scrolling area][] will
extend rightward or downward to include it. In fact, if the ancestor is
`overflow:scroll` or `auto`, you can even "physically" (like, with your
mouse, or let's be realistic here, your trackpad) scroll down or
rightward to see the entirety of whatever `position:absolute` thing is
sticking out---which is quite confusing in this case since it's
invisible. #NewAndExcitingWaysToIntroduceMysteriousWhitespaceWithCSS

[scrolling area]: http://dev.w3.org/csswg/cssom-view/#scrolling-area

Why in the world does the scrolling area do that, and only rightward and
downward but not leftward or upward? #ScienceCannotExplain
(Fun fact, it can instead extend leftward but not rightward, if the
"writing mode" is configured for right-to-left text for languages like
Arabic. With sufficiently confusing "writing mode" settings it can even
extend upward but not downard---and no, it's not just "down-to-up" text,
there's no such thing. Feel free to read the spec I linked to above.)

--

I wanted this to be an automated rather than visual test but there's no
obvious suite this goes in, so I'm creating a new 'CSS' unit test suite
for automated visual tests (hence adding the stylesheet 2 commits ago).
There are automated tests in `test/visual.html` that should probably be
moved here.