Fix selection in numerator overlaps fraction bar
Not just selection of course, descenders and LatexCommandInput were also
visibly overlapping with the fraction bar.

The original reason for the negative margin was the numerator was too
far from the fraction bar, and the denominator too close:
0b754f1103eea198be641410ce0d7e9c0174e1b6

Now instead we remove the negative margin but push the denominator even
further from the fraction bar to preserve their symmetrical distance.