Fix paren vert-align at fractional font-sizes in Firefox
Apparently latest Firefox (34) doesn't round DOM measurements to the
nearest pixel (anymore?), so forcing rounding in the denominator leads
to noticeable misalignment of parentheses. Firefox also doesn't round
`getComputedStyle(this.contentjQ[0]).height`, but that stuff is annoying
because it's different in IE8 and maybe Firefox <4 or something[1], and
rounding errors in the numerator are no big deal, the denominator is
always >1px so rounding error in the numerator can only result in errors
of <1px. (Whereas at a small font-size like 14px, rounding error can be
a very noticeable >7% of the total height of the contents of parens.)

[1]: https://developer.mozilla.org/en-US/docs/Web/API/window.getComputedStyle