No cursor.show().replaceSelection() before .write()
Change the contract of .write() on Node's so that the caller is no
longer supposed to ensure the cursor is shown and pass in the selection
as `replacedFragment`. Now instead it's up to .write() itself to do
`cursor.show().replaceSelection()` (as MathBlock::write does), or do
`cursor.show().deleteSelection()` (as all the places that used to do
`replacedFragment.remove()` now do), or even to be a noop (which the
next commit on feature.autoSubscriptNumeral will have).