Add 'autoSubscriptNumerals' option
When enabled, typing `x1` results in `x_1` with the cursor to the right
of the subscript. Typing more numerals extends the subscript without
putting the cursor inside the subscript (thanks `contactWeld`), and
backspacing "peels off" the last symbol in the subscript. Left and right
arrow keys go around the subscript, skipping its contents entirely. The
down arrow still goes into the subscript, but inside the subscript,
typing _ is ignored, and typing anything other than a symbol (^ or /)
pretends the cursor is to the right of the subscript.

Note that it's still possible for a non-Symbol command to be in a
subscript, by pasting, leading to the edge case of what to do when
backspacing a subscript whose last command is not a symbol; I elected to
inject the cursor at the end of the subscript right after the last
command and backspace it (which should select it, and another backspace
should delete it, just like if it hadn't been in the subscript).

Also note that in order to properly ignore typing _ inside a subscript,
the previous commit was necessary, or else selecting something in the
subscript and then typing _ would detach the selection. Relatedly, a
tiny refactor of MathBlock::write() is necessary so that `.write()` on
the subscript block can inherit the right things.