Enforce sub-then-sup order when deleting SupSub blocks
Say you've got

    x_{sub}^{|sup}

where the | is the cursor. If you hit backspace, what should you get?

    x_{sub}|sup

right?

Well, prior to this commit, you'd get

    x|sup_{sub}

which maybe would make sense if subscripts and superscripts were
semantically on equal footing (e.g. Einstein notation or chemical
symbols symbol), but usually subscripts are part of a variable's name
(e.g. the index), whereas superscripts are exponents.

We already enforce the subscript-then-superscript ordering when
exporting LaTeX and when moving the cursor around with the left and
right arrow keys (when `leftRightIntoCmdGoes` isn't set), so it makes
sense to be consistent with that when deleting.