Fix backspacing through autosubscripts due to merge
Thanks to #434, if you have both a subscript and an exponent (x_a^b),
deleting out of the subscript in any direction puts you left of the
exponent, which means the `-dir` fails to prevent the cursor from going
"through" the exponent when backspacing an autosubscript.

So now after deleting the subscript, insert the cursor on the `-dir`
side we want it to be at.

Note that the insertion is invalid if there wasn't an exponent in
addition to the subscript, because if there was just a subscript, the
whole SupSub node gets removed. But in that case we don't need to insert
the cursor anywhere, it's already in the right place.

In the 'backspace through subscript & superscript' test, note that the
"slightly weird behavior" is exactly what #434 fixes.