`edit`, `edited` as aliases for `reflow` event
What is currently on the `dev` branch known as the `reflow` event was
called `edited` when it was introduced in #241. @alopatin then brought
up a serious issue (#317) with the `edited` event as it was then
implemented, which is that you could edit < into ≤ (etc) without
triggering the event.

Rather than fixing the problem, I obstinately renamed the event from
`edited` to `reflow` so that I could claim that the event not being
triggered when no sizes change was a feature, not a bug (#325).

@jwmerrill finally cut through my bullshit and got @alopatin's fix
from #317 merged into `dev` (tweaked as #425), which made edits like
< into ≤ call the `reflow` handler, so that there'd be at least some
handler that's reliably called when the contents of a field changes.

Because the `reflow` event was so broken for so long, people like
@alopatin have been using a fork of `dev` with a fixed `edited`.
To ease their migration to v0.10.0, and also because the name makes
more sense for its current role, I want to add back `edited` as an
alias for `reflow` (without removing `reflow`).

While we're talking about a name that makes sense, what name for this
event do we want to recommend going forward? How about we get rid of
the inexplicable past tense and just call the event `edit`. Make that
the recommended alias in the docs.