Refactor API obj constructors to take in controller
More prep for more than one API object per controller, so that
eventually we could just call the API constructor on the controller
of an existing MathQuill instance and get a new API object for it,
without mutating the existing instance or its controller at all,
simply hooking itself up to it.

For that to work, when for example the `.MathField()` API function is
MathQuill-ifying an element into a math field, it has to first create
the root block and controller to pass it to the API constructor, which
merely hooks stuff up, then another method has to be called to actually
MathQuill-ify of the element into a math field, `.__mathquillify()`.

Also, now that inheriting `init` actually matters, its important that
MathField and TextField inherit from the actual EditableField API class
and not just the EditableField API function with the same prototype, so
now there's an `EditableField` API class separate from the
`MQ.EditableField` API function.

Finally, for inner math fields (`\MathQuillMathField{}`), there's no API
function call, we'll just have MathQuillMathField::finalizeTree() create
the controller etc, including a `MathField` instance (the API class, not
the `MQ.MathField` API function).