Add back \lbrack and \rbrack, but as plain symbols
...that are just aliases for the '[' and ']' vanilla symbols.

Before #262, \lbrack and \rbrack were parsed in the same totally broken
way that \lparen, \lbrace etc were.

\lparen is not a real LaTeX command and since it was so broken I highly
doubt anyone was relying on how it was parsed, so it's okay that it was
removed in #262.

However, unlike \lparen but like \lbrace, \lbrack and \rbrack are real
LaTeX commands [1]. So, like we did for \lbrace and \rbrace in 38c2b09
on #323, add vanilla symbols for them.

[1]: http://tex.stackexchange.com/a/45984
also, http://tex.stackexchange.com/a/294104