Update comment on LaTeX math parsers
The technique of parsers that are "supposed to" yield MathCommands
instead yielding a Fragment of MathCommands (which can be transparently
substituted for them when inserting into a MathBlock) apparently dates
back to LatexFragment::parser():

https://github.com/mathquill/mathquill/commit/e4ebf64bcedab76022bc417513403a4d1d7a20ef#diff-58a03b51243f0041858a5394b0a40090R153

And has since also been used OperatorName::parser() and
LatexCmds.operatorname::parser().

I had to figure out for myself all over again how it works and was
mislead by this comment in the LaTeX math parser, so, update it.