Fix ARIA re-announcing the last thing over again
For example, if you type 'x', it announces 'x'; if you backspace, it
should announce 'x' again, but I was finding in Safari 9.1.1 with
VoiceOver on Mac OS X 10.11.5 that the second time it wouldn't be
announced. (This also comes if you have 22 and you're moving your cursor
through them, 'two' only gets announced once; or if you Left arrow over
an 'x' then Right arrow back; or lots of other scenarios. This is bad.)

I'm able to reproduce with my Safari + VoiceOver that our old
.empty().html('...') trick in isolation no longer works:
http://jsbin.com/rekinorami/edit?html,js,output

To my surprise, setTimeout() didn't fix it. However, replacing .empty()
with .html(' '), which sets the contents to something that is nonempty
but is empty whitespace so it doesn't announce anything, does work.
Hope this trick lasts?! @sclower, can you give it a spin in some more
screen readers, or do you have any better ideas to future-proof this?